+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- UMD build format from Rollup config
+
 ## [1.5.0] - 2025-07-07
 
 ### Added